Introduces Paxos, State Machines and a new version of High Availability that depends on those for master election
* Introduces the cluster component, which encapsulates all cluster
   related functionality, for creating and running connected state
   machines, implementing paxos, timeouts, heartbeats and delivering
   events around those
* Drastically changes the design of HA to use the above, removing
   ZooKeeper completely. Highly Available GraphDatabase is broken
   down into smaller components that listen on events from the clustering
   layer and react accordingly, taking on master or slave roles.
* The mode switching of master/slave is performed in state machines,
   allowing for intermediate states to exist, broadcasting events as
   they happen
* Backup is updated to work with those changes
*  JMX beans are updated to provide information relevant to the new
   state of things

Authors: Anders Nawroth, Johan Svensson, Mattias Persson, Rickard Oberg, Chris Gioran

+++ enterprise/com/src/main/java/org/neo4j/com/DechunkingChannelBuffer.java
@@ -44,12 +44,12 @@ public class DechunkingChannelBuffer implements ChannelBuffer
     private ChannelBuffer buffer;
     private boolean more;
     private boolean hasMarkedReaderIndex;
-    private final int timeoutSeconds;
+    private final long timeoutSeconds;
     private boolean failure;
     private final byte applicationProtocolVersion;
     private final byte internalProtocolVersion;
 
-    DechunkingChannelBuffer( BlockingReadHandler<ChannelBuffer> reader, int timeoutSeconds, byte internalProtocolVersion,
+    DechunkingChannelBuffer( BlockingReadHandler<ChannelBuffer> reader, long timeoutSeconds, byte internalProtocolVersion,
             byte applicationProtocolVersion )
     {
         this.reader = reader;
